Added Drum Weapon
Seismic slam currently has bug where it instantly pops bc it thinks it's touching a wall when the player is sort of close to a wall. This should be fixed in the future once we have art for the seismic slam projectile and we don't need to stretch the transform like I'm doing now.